alligning file structure
